+++ rotkehlchen/chain/ethereum/modules/curve/balances.py
@@ -10,7 +10,6 @@ from .constants import CPT_CURVE
 if TYPE_CHECKING:
     from rotkehlchen.chain.evm.node_inquirer import EvmNodeInquirer
     from rotkehlchen.history.events.structures.evm_event import EvmEvent
-    from rotkehlchen.types import CHAIN_IDS_WITH_BALANCE_PROTOCOLS
 
 
 class CurveBalances(ProtocolWithGauges):
@@ -19,15 +18,10 @@ class CurveBalances(ProtocolWithGauges):
     LP tokens are already queried by the normal token detection.
     """
 
-    def __init__(
-            self, database: DBHandler,
-            evm_inquirer: 'EvmNodeInquirer',
-            chain_id: 'CHAIN_IDS_WITH_BALANCE_PROTOCOLS',
-    ):
+    def __init__(self, database: DBHandler, evm_inquirer: 'EvmNodeInquirer'):
         super().__init__(
             database=database,
             evm_inquirer=evm_inquirer,
-            chain_id=chain_id,
             counterparty=CPT_CURVE,
             deposit_event_types={(HistoryEventType.DEPOSIT, HistoryEventSubType.DEPOSIT_ASSET)},
             gauge_deposit_event_types={(HistoryEventType.DEPOSIT, HistoryEventSubType.DEPOSIT_ASSET)},  # noqa: E501