+++ rotkehlchen/chain/optimism/modules/velodrome/balances.py
@@ -6,7 +6,6 @@ from rotkehlchen.db.dbhandler import DBHandler
 
 if TYPE_CHECKING:
     from rotkehlchen.chain.optimism.node_inquirer import OptimismInquirer
-    from rotkehlchen.types import CHAIN_IDS_WITH_BALANCE_PROTOCOLS
 
 
 class VelodromeBalances(VelodromeLikeBalances):
@@ -14,11 +13,9 @@ class VelodromeBalances(VelodromeLikeBalances):
             self,
             database: DBHandler,
             evm_inquirer: 'OptimismInquirer',
-            chain_id: 'CHAIN_IDS_WITH_BALANCE_PROTOCOLS',
     ):
         super().__init__(
             database=database,
             evm_inquirer=evm_inquirer,
-            chain_id=chain_id,
             counterparty=CPT_VELODROME,
         )